requirements.txt
selenium
pytest
pytest-html
allure-pytest
behave
jsonschema
PyYAML
requests
webdriver-manager